bennett-kew-report-automate/requirements.txt
anthropic>=0.40.0
pymupdf>=1.25.0
pillow>=10.0.0
python-dotenv>=1.0.0
msal>=1.28.0
requests>=2.31.0
